Add EUI64 addresses support
EUI64 addresses are common for multiple protocols.  This commit adds
support for it in general fields.

It is currently used in Zigbee.

The decision of the internal representation of the EUI64 addresses is
copied from IPField and IP6Field: a human-readable string.  Warning: old
code expecting integer-type addresses (e.g. tests) has to be fixed.